Create Authentication Plugins
Provides the framework for creating authentication plugins and using
them from a session object.

To allow this system to co-exist with the original client there is a bit
of a hack. The client object itself is now also an authentication
plugin, that supports the original client pattern. If a client is
created without a session object then that session object uses the
client as it's authentication plugin.

What this means is that when calling client.authenticate we hand off to
the session.authenticate function to authenticate the session object.
In the original case this will then call back into the client and
authentication will proceed as normal.

If someone calls client.authenticate and there is another plugin in it's
place then the client will take the auth_ref from that plugin and
attempt to fill the client variables in for compatibility.

Similarly if the session prompts a re-authenticate then it will call
back to the client to update it's token and maintain compatibility.

Change-Id: I682c8dcd3705148aaa804a91f4ed48a5b74bdc12
blueprint: auth-plugins